-Updated the Business details section with comments explaining different lines of code and functionality.
-Did the same for the business index as well.

-Added a comment card page with different funcitonality as such: Component Structure:

The component is a simple card layout, wrapped in a div with the class "border col-sm-4".
Review Type Display:

The component displays a heading (h2 element) with the class "rant". The content of this heading depends on whether the comment.rant property is truthy or falsy. If comment.rant is truthy, it shows bad or good reviews.
Comment Content:

The component displays the comment content using an h4 element with the content of comment.content.
Author Information:

The author's username is displayed in a strong element within an h3 element, formatted as "- {comment.author.userName}".
Rating Information:

The component displays the comment's rating using an h4 element with the content "Rating: {comment.stars}".
Delete Button:

A delete button is included with the text "Delete Comment". When clicked, it triggers the onDelete function passed as a prop.